Pivoting to standard. Removal of CitizenVotingAnalytics component and its tracking logic, replacing its functionality in VotingActions and VotingColumn with direct use of the Analytics provider. Updated VotingSidebar to include page view tracking.
