+++ app/src/components/proposals/proposalPage/VotingSidebar/VotingSidebar.tsx
@@ -1,4 +1,5 @@
-import React from "react"
+"use client"
+import React, { useEffect, useRef } from "react"
 
 import {
   VotingColumnProps,
@@ -9,12 +10,17 @@ import VotingCard, {
 } from "@/components/proposals/proposalPage/VotingSidebar/votingCard/VotingCard"
 import VotingColumn from "@/components/proposals/proposalPage/VotingSidebar/votingColumn/VotingColumn"
 import VotingRedirect from "@/components/proposals/proposalPage/VotingSidebar/VotingRedirect"
+import { useAnalytics } from "@/providers/AnalyticsProvider"
 
 interface VotingSidebarProps {
   className?: string
   votingCardProps: VotingCardProps
   votingColumnProps: VotingColumnProps
   votingRedirectProps?: VotingRedirectProps
+  proposalId: string
+  citizen?: any
+  citizenEligibility?: any
+  proposalType?: string
 }
 
 const VotingSidebar = ({
@@ -22,16 +28,55 @@ const VotingSidebar = ({
   votingCardProps,
   votingColumnProps,
   votingRedirectProps,
-}: VotingSidebarProps) => (
-  <div className={`w-[304px] gap-6 flex flex-col ${className}`}>
-    <div className="w-[304px] ">
-      <VotingCard {...votingCardProps} />
-      <VotingColumn {...votingColumnProps} />
-      <div className="mt-5">
-        {votingRedirectProps && <VotingRedirect {...votingRedirectProps} />}
+  proposalId,
+  citizen,
+  citizenEligibility,
+  proposalType,
+}: VotingSidebarProps) => {
+  const { track } = useAnalytics()
+  const isTracked = useRef(false)
+
+  // Determine citizen status
+  const getCitizenStatus = () => {
+    if (citizen) {
+      return "registered"
+    } else if (citizenEligibility?.eligible) {
+      return "eligible"
+    } else {
+      return "ineligible"
+    }
+  }
+
+  useEffect(() => {
+    if (!isTracked.current) {
+      // Page View event - when user visits the voting page
+      track("Citizen Voting Page View", {
+        page_title: "Proposal Voting",
+        page_type: proposalType,
+        proposal_id: proposalId,
+        citizen_status: getCitizenStatus(),
+      })
+      isTracked.current = true
+    }
+  }, [
+    proposalId,
+    citizen,
+    citizenEligibility,
+    proposalType,
+    track,
+  ])
+
+  return (
+    <div className={`w-[304px] gap-6 flex flex-col ${className}`}>
+      <div className="w-[304px] ">
+        <VotingCard {...votingCardProps} />
+        <VotingColumn {...votingColumnProps} />
+        <div className="mt-5">
+          {votingRedirectProps && <VotingRedirect {...votingRedirectProps} />}
+        </div>
       </div>
     </div>
-  </div>
-)
+  )
+}
 
 export default VotingSidebar
